implement nav flag switch
